requirements.txt
aiogram==3.15.0
beautifulsoup4==4.9.1
python-dotenv==1.0.1
pytonconnect==0.3.1
pytoniq_core==0.1.40
redis==5.2.0
Requests==2.32.3
selenium==4.27.1
YooMoney==0.1.1
